requirements.txt
asgiref==3.2.7
certifi==2019.11.28
cycler==0.10.0
Django==3.0.4
intrinio-sdk==5.1.0
kiwisolver==1.1.0
matplotlib==3.2.1
numpy==1.18.2
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
